+++ evals/tasks/t-001-add-sum-function/solution/utils.py
@@ -1,5 +1,0 @@
-"""Utility module."""
-
-
-def sum_list(xs):
-    return sum(xs)