+++ evals/tasks/t-003-rename-function/starter/app.py
@@ -1,9 +1,0 @@
-from greeting import old_greeting
-
-
-def greet_world():
-    return old_greeting("world")
-
-
-def greet_team():
-    return old_greeting("team")